+++ projects/angular-sdk/src/i18n/en.ts
@@ -278,7 +278,6 @@ export const locale: I18n = {
         },
         CFAR_OFFER_DIALOG: {
             TITLE: "Cancel For Any Reason",
-            POWERED_BY: "Powered by Hopper",
             VIEW_DETAILS: "View details",
             BUTTON: {
                 CANCEL: "Continue Without Coverage",