+++ projects/angular-sdk/src/i18n/es.ts
@@ -278,7 +278,6 @@ export const locale: I18n = {
         },
         CFAR_OFFER_DIALOG: {
             TITLE: "Cancelar por cualquier motivo",
-            POWERED_BY: "Impulsado por Hopper",
             VIEW_DETAILS: "Ver detalles",
             BUTTON: {
                 CANCEL: "Continuar sin cobertura",