+++ projects/angular-sdk/src/i18n/i18n.interface.ts
@@ -276,7 +276,6 @@ export interface I18n {
         },
         CFAR_OFFER_DIALOG: {
             TITLE: string;
-            POWERED_BY: string;
             VIEW_DETAILS: string;
             BUTTON: {
                 CANCEL: string;