+++ projects/angular-sdk/src/i18n/i18n.interface.ts
@@ -255,7 +255,8 @@ export interface I18n {
             },
             BUTTON: {
                 CLOSE: string;
-            }
+            },
+            TERMS_LINK_TEXT: string;
         },
         CFAR_OFFER_BANNER: {
             TITLE: string;
@@ -284,7 +285,6 @@ export interface I18n {
             NO_OFFERS: string;
             LOADING: string;
             TERMS_WARNING: string;
-            TERMS_LINK_TEXT: string;
         },
         CFAR_EXERCISE_DIALOG: {
             TITLE: string;